backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
langchain
langchain-openai
openai
pypdf
pypdf2
python-docx
matplotlib
pandas
openpyxl
python-dotenv
streamlit
pydantic-settings
faiss-cpu
sentence_transformers
langchain_community
# sqlmodel
mysql-connector-python
pymysql

-e .
